Ship a flag file; switch to Makefile
Because someone might uninstall but not purge dconf-config, we
should ensure the package is installed before executing the
Xsession.d code.  Unfortunately, the dconf profile is also a
conffile.  So we ship a flag file, and test for that, because
running dpkg-query takes forever on -cluster

Also, switch to a Makefile and let dh_auto_foo deal, rather than
an override in dh_install